backend/requirements.txt
crawl4ai
fastapi[standard]
google-genai
pydantic
python-dotenv
